Refactor common code
